Remove unnecessary interfaces and thin wrappers
- Delete BotEngine interface → rename BotEngineImpl to BotEngine
- Delete AuthService interface → TelegramAuthService used directly
- Delete TextRenderer interface → TelegramRenderer used directly
- Delete MessengerNotificationService interface → TelegramNotificationService used directly
- Delete TelegramTextService thin wrapper → TelegramAdapter uses TelegramRenderer
- Update MainMenuFactory, BotEngine, test to use TelegramAuthService directly
