+++ chatbot/src/main/java/pt/estga/chatbot/features/core/MainMenuFactory.java
@@ -6,27 +6,24 @@ import pt.estga.chatbot.config.ChatbotAuthProperties;
 import pt.estga.chatbot.constants.CallbackData;
 import pt.estga.chatbot.constants.MessageKey;
 import pt.estga.chatbot.models.BotInput;
-import pt.estga.chatbot.models.Platform;
 import pt.estga.chatbot.models.ui.Button;
 import pt.estga.chatbot.models.ui.Menu;
-import pt.estga.chatbot.services.AuthService;
 import pt.estga.chatbot.services.messages.UiTextService;
+import pt.estga.chatbot.telegram.services.TelegramAuthService;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @Component
 @RequiredArgsConstructor
 public class MainMenuFactory {
 
-    private final List<AuthService> authServices;
+    private final TelegramAuthService telegramAuthService;
     private final UiTextService textService;
     private final ChatbotAuthProperties chatbotAuthProperties;
 
     public Menu create(BotInput input) {
-        AuthService authService = resolveAuthService(input.getPlatform());
-        boolean isAuthenticated = authService.isAuthenticated(input.getUserId());
+        boolean isAuthenticated = telegramAuthService.isAuthenticated(input.getUserId());
 
         List<List<Button>> buttonRows = new ArrayList<>();
         boolean canStartSubmission = chatbotAuthProperties.isOptional() || isAuthenticated;
@@ -53,12 +50,5 @@ public class MainMenuFactory {
                 .titleNode(textService.get(MessageKey.HELP_OPTIONS_TITLE))
                 .buttons(buttonRows)
                 .build();
-    }
-
-    private AuthService resolveAuthService(Platform platform) {
-        return authServices.stream()
-                .filter(s -> s.supports(platform))
-                .findFirst()
-                .orElseThrow(() -> new NoSuchElementException("No AuthService found for platform: " + platform));
     }
 }
