--- chatbot/src/main/java/pt/estga/chatbot/services/AuthService.java
@@ -1,16 +1,0 @@
-package pt.estga.chatbot.services;
-
-import pt.estga.chatbot.models.Platform;
-import pt.estga.commoncore.models.AppPrincipal;
-
-import java.util.Optional;
-
-public interface AuthService {
-
-    boolean isAuthenticated(String platformUserId);
-
-    Optional<AppPrincipal> authenticate(String platformUserId);
-
-    boolean supports(Platform platform);
-
-}
