--- chatbot/src/main/java/pt/estga/chatbot/services/messages/TextRenderer.java
@@ -1,8 +1,0 @@
-package pt.estga.chatbot.services.messages;
-
-import pt.estga.chatbot.models.text.RenderedText;
-import pt.estga.chatbot.models.text.RichText;
-
-public interface TextRenderer {
-    RenderedText render(RichText node);
-}
