--- chatbot/src/main/java/pt/estga/chatbot/services/notifications/MessengerNotificationService.java
@@ -1,18 +1,0 @@
-package pt.estga.chatbot.services.notifications;
-
-import pt.estga.chatbot.constants.MessageKey;
-import pt.estga.chatbot.models.Message;
-import pt.estga.chatbot.models.Platform;
-
-public interface MessengerNotificationService {
-
-    boolean supports(Platform platform);
-
-    void sendNotification(String recipientId, Message message);
-
-    void sendNotification(String recipientId, MessageKey messageKey, Object... args);
-
-    void sendNotificationWithMenu(String recipientId, Message message);
-
-    void sendNotificationWithMenu(String recipientId, MessageKey messageKey, Object... args);
-}
