+++ chatbot/src/main/java/pt/estga/chatbot/telegram/services/TelegramAuthService.java
@@ -4,8 +4,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
-import pt.estga.chatbot.models.Platform;
-import pt.estga.chatbot.services.AuthService;
 import pt.estga.commoncore.models.AppPrincipal;
 import pt.estga.user.entities.User;
 import pt.estga.user.repositories.UserRepository;
@@ -16,16 +14,14 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class TelegramAuthService implements AuthService {
+public class TelegramAuthService {
 
     private final UserRepository userRepository;
 
-    @Override
     public boolean isAuthenticated(String platformUserId) {
         return userRepository.findByTelegramChatId(platformUserId).isPresent();
     }
 
-    @Override
     public Optional<AppPrincipal> authenticate(String platformUserId) {
         return userRepository.findByTelegramChatId(platformUserId)
                 .map(this::toPrincipal);
@@ -42,10 +38,5 @@ public class TelegramAuthService implements AuthService {
                 .enabled(user.isEnabled())
                 .accountNonLocked(!user.isAccountLocked())
                 .build();
-    }
-
-    @Override
-    public boolean supports(Platform platform) {
-        return Platform.TELEGRAM.equals(platform);
     }
 }
